test: use separate Playwright instance to automate inspector
This will prevent inspector from picking up test actions.
